Ammo counter, health bar, player damage, zombie attack animation
Added the ammo counter. Didn't know where to put it so I just attached
it to the player.
Added a health bar. Didn;t know where to put it so I just attached it to
the player. It looks really stupid and needs to be moved.
Tested player damage and zombie attack animations with one zombie. In
order to get the animations to be relative, the zombie sprite had to be
a child of another game object. It's the one with the triangle
reenabled. Try and aggro just that one. He kind of does this lungey
thing.
Players take damage when they collide with a zombie. Nothing happens
when they fall below 0 health yet though.